requirements.txt
Django==4.2.13
django-environ==0.11.2
mysqlclient==2.2.4
pillow==10.3.0
djangorestframework==3.15.1
markdown==3.6
django-filter==24.2
pymemcache==4.0.0
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0